selectors: merge back company & job-title
